Trying to implement searching with scopes
Thus far, a person can search for a carrier by name. For example,
if you searched for a carrier with the word "Beco" in it, you could
go http://localhost:3000/carriers?starts_with=Beco to see the results.
Currently working through understanding how to set up the search form
in the view correctly.